Revert schema files and edit them manually
It appears that many people are having issues with using the tool as a result of my changes. I am fairly certain the issue has to do with the regenerated schemas done after editing store_download_req.json. This commit reverts them, and instead uses find+replace to manually edit the references to use externalVersionId instead of appExtVrsId.